productService update
