+++ src/main/java/com/ohgiraffers/warehousemanagement/wms/product/controller/ProductController.java
@@ -4,6 +4,8 @@ import com.ohgiraffers.warehousemanagement.wms.product.model.DTO.ProductCreateDT
 import com.ohgiraffers.warehousemanagement.wms.product.model.DTO.ProductPageResponseDTO;
 import com.ohgiraffers.warehousemanagement.wms.product.model.DTO.ProductResponseDTO;
 import com.ohgiraffers.warehousemanagement.wms.product.service.ProductService;
+import com.ohgiraffers.warehousemanagement.wms.product.service.ProductServiceImpl;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -12,24 +14,23 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/products")
 public class ProductController {
 
+    private final ProductServiceImpl productServiceImpl;
     private final ProductService productService;
 
-    public ProductController(ProductService productService) {
+    @Autowired
+    public ProductController(ProductServiceImpl productServiceImpl, ProductService productService) {
+        this.productServiceImpl = productServiceImpl;
         this.productService = productService;
     }
 
     // 상품 목록 조회
-    @GetMapping("")
+    @GetMapping
     public String getAllProducts(
-            @RequestParam(value = "page", defaultValue = "1") int page,
-            @RequestParam(value = "searchKeyword", required = false) String searchKeyword,
+            @RequestParam(defaultValue = "1") int page,
+            @RequestParam(defaultValue = "10") int pageSize,
+            @RequestParam(required = false) String searchKeyword,
             Model model) {
-        int pageSize = 10;
-        ProductPageResponseDTO productPage = productService.getAllProducts(page, pageSize, searchKeyword);
-
-        model.addAttribute("pageTitle", "상품 관리");
-        model.addAttribute("cardTitle", "상품 목록");
-        model.addAttribute("cardDescription", "등록된 모든 상품을 확인하고 관리할 수 있습니다.");
+        ProductPageResponseDTO productPage = productServiceImpl.getAllProducts(page, pageSize, searchKeyword);
         model.addAttribute("products", productPage.getProducts());
         model.addAttribute("currentPage", productPage.getCurrentPage());
         model.addAttribute("totalPages", productPage.getTotalPages());
@@ -37,69 +38,53 @@ public class ProductController {
         model.addAttribute("startItem", productPage.getStartItem());
         model.addAttribute("endItem", productPage.getEndItem());
         model.addAttribute("searchKeyword", searchKeyword);
-
+        model.addAttribute("pageTitle", "상품 관리");
+        model.addAttribute("cardTitle", "상품 목록");
+        model.addAttribute("cardDescription", "등록된 모든 상품을 확인하고 관리할 수 있습니다.");
         return "products/list";
     }
 
-    // 특정 상품 조회 (수정 폼)
-    @GetMapping("/{id}")
-    public String getProductById(@PathVariable("id") Integer id, Model model) {
-        ProductResponseDTO product = productService.getProductById(id);
-        model.addAttribute("pageTitle", "상품 수정 - " + product.getProductName());
-        model.addAttribute("product", product);
-        model.addAttribute("categories", productService.getCategories());
-        return "products/update";
-    }
-
-    // 신규 상품 등록 폼 표시
+    // 상품 생성 페이지
     @GetMapping("/create")
-    public String showCreateProductForm(Model model) {
-        model.addAttribute("pageTitle", "신규 상품 등록");
+    public String createProductForm(Model model) {
         model.addAttribute("product", new ProductCreateDTO());
-        model.addAttribute("categories", productService.getCategories());
+        model.addAttribute("categories", productServiceImpl.getCategories());
         return "products/create";
     }
 
     // 상품 생성
     @PostMapping("/create")
-    public String createProduct(@ModelAttribute ProductCreateDTO createDTO, Model model) {
+    public String createProduct(@ModelAttribute("product") ProductCreateDTO productCreateDTO, Model model) {
         try {
-            productService.createProduct(createDTO);
+            productServiceImpl.createProduct(productCreateDTO);
             return "redirect:/products";
         } catch (IllegalArgumentException e) {
             model.addAttribute("errorMessage", e.getMessage());
-            model.addAttribute("pageTitle", "신규 상품 등록");
-            model.addAttribute("product", createDTO);
-            model.addAttribute("categories", productService.getCategories());
+            model.addAttribute("categories", productServiceImpl.getCategories());
             return "products/create";
         }
     }
 
+    // 상품 수정 페이지
+    @GetMapping("/{id}")
+    public String updateProductForm(@PathVariable("id") Integer id, Model model) {
+        ProductResponseDTO product = productService.getProductById(id);
+        model.addAttribute("product", product);
+        model.addAttribute("categories", productServiceImpl.getCategories());
+        return "products/update";
+    }
+
     // 상품 수정
     @PostMapping("/update/{id}")
-    public String updateProduct(@PathVariable("id") Integer id, @ModelAttribute ProductCreateDTO updateDTO, Model model) {
-        try {
-            productService.updateProduct(id, updateDTO);
-            return "redirect:/products";
-        } catch (IllegalArgumentException e) {
-            model.addAttribute("errorMessage", e.getMessage());
-            model.addAttribute("pageTitle", "상품 수정 - " + updateDTO.getProductName());
-            model.addAttribute("product", updateDTO);
-            model.addAttribute("categories", productService.getCategories());
-            return "products/update";
-        }
+    public String updateProduct(@PathVariable("id") Integer id, @ModelAttribute("product") ProductCreateDTO productCreateDTO) {
+        productServiceImpl.updateProduct(id, productCreateDTO);
+        return "redirect:/products";
     }
 
     // 상품 삭제
     @PostMapping("/delete/{id}")
     public String deleteProduct(@PathVariable("id") Integer id) {
-        productService.deleteProduct(id);
-        return "redirect:/products";
-    }
-
-    // 상품 내보내기 (임시 리다이렉트)
-    @GetMapping("/export")
-    public String exportProducts() {
+        productServiceImpl.deleteProduct(id);
         return "redirect:/products";
     }
 }
